+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Underscore.php
 - Intial release of Underscore.php
 - Type classes are now extendable
 - Macros can't conflict between types
+- Added Arrays::replaceValue to do an str_replace